test: fix testcase
